sentiment_project/requirements.txt
Django==4.2.19
djangorestframework==3.14.0
pandas==1.5.3
scikit-learn==1.0.2
scipy==1.7.3
joblib==1.2.0